+++ src/drivers/lwip_unsupported.c
@@ -1,4 +1,4 @@
-// SPDX-FileCopyrightText: (c) 2021-2023 Shawn Silverman <shawn@pobox.com>
+// SPDX-FileCopyrightText: (c) 2021-2024 Shawn Silverman <shawn@pobox.com>
 // SPDX-License-Identifier: AGPL-3.0-or-later
 
 // lwip_unsupported.c contains the unsupported Ethernet interface implementation.
@@ -8,48 +8,50 @@
 
 #include "lwip_driver.h"
 
-bool enet_is_unknown() {
+bool driver_is_unknown() {
   return false;
 }
 
-void enet_get_mac(uint8_t *mac) {
+void driver_get_system_mac(uint8_t *mac) {
 }
 
-bool enet_has_hardware() {
+void driver_set_mac(uint8_t mac[ETH_HWADDR_LEN]) {
+}
+
+bool driver_has_hardware() {
   return false;
 }
 
-bool enet_init(const uint8_t mac[ETH_HWADDR_LEN],
-               netif_ext_callback_fn callback) {
+bool driver_init(const uint8_t mac[ETH_HWADDR_LEN]) {
   return false;
 }
 
-void enet_deinit() {
+void driver_deinit() {
 }
 
-struct netif *enet_netif() {
-  return NULL;
+void driver_proc_input(struct netif *netif) {
 }
 
-void enet_proc_input(void) {
+void driver_poll(struct netif *netif) {
 }
 
-void enet_poll() {
+err_t driver_output(struct netif *netif, struct pbuf *p) {
+  return ERR_IF;
 }
 
-int phy_link_speed() {
+int driver_link_speed() {
   return 0;
 }
 
-bool phy_link_is_full_duplex() {
+bool driver_link_is_full_duplex() {
   return false;
 }
 
-bool phy_link_is_crossover() {
+bool driver_link_is_crossover() {
   return false;
 }
 
-bool enet_output_frame(const uint8_t *frame, size_t len) {
+bool driver_output_frame(const uint8_t *frame, size_t len) {
   return false;
 }
 
@@ -59,15 +61,7 @@ bool enet_output_frame(const uint8_t *frame, size_t len) {
 
 #if !QNETHERNET_ENABLE_PROMISCUOUS_MODE
 
-bool enet_set_mac_address_allowed(const uint8_t *mac, bool allow) {
-  return false;
-}
-
-bool enet_join_group(const ip4_addr_t *group) {
-  return false;
-}
-
-bool enet_leave_group(const ip4_addr_t *group) {
+bool driver_set_mac_address_allowed(const uint8_t *mac, bool allow) {
   return false;
 }