+++ test/test_ethernet/test_main.cpp
@@ -244,8 +244,8 @@ static void test_other_null_mac() {
   Ethernet.begin(nullptr, INADDR_NONE, INADDR_NONE, INADDR_NONE, INADDR_NONE);
 #pragma GCC diagnostic pop
 
-  TEST_ASSERT_FALSE(enet_set_mac_address_allowed(nullptr, true));
-  TEST_ASSERT_FALSE(enet_set_mac_address_allowed(nullptr, false));
+  TEST_ASSERT_FALSE(driver_set_mac_address_allowed(nullptr, true));
+  TEST_ASSERT_FALSE(driver_set_mac_address_allowed(nullptr, false));
 }
 
 // Tests NULL join/leave groups.